refactor(native-agent): extract reusable runner
